+++ packages/vite/src/node/ssr/ssrTransform.ts
@@ -94,7 +94,7 @@ async function ssrTransformScript(
   function defineImport(node: Node, source: string) {
     deps.add(source)
     const importId = `__vite_ssr_import_${uid++}__`
-    s.appendLeft(
+    s.appendRight(
       node.start,
       `const ${importId} = await ${ssrImportKey}(${JSON.stringify(source)});\n`
     )
@@ -115,6 +115,7 @@ async function ssrTransformScript(
     // import { baz } from 'foo' --> baz -> __import_foo__.baz
     // import * as ok from 'foo' --> ok -> __import_foo__
     if (node.type === 'ImportDeclaration') {
+      s.remove(node.start, node.end)
       const importId = defineImport(node, node.source.value as string)
       for (const spec of node.specifiers) {
         if (spec.type === 'ImportSpecifier') {
@@ -129,7 +130,6 @@ async function ssrTransformScript(
           idToImportMap.set(spec.local.name, importId)
         }
       }
-      s.remove(node.start, node.end)
     }
   }
 
@@ -207,13 +207,11 @@ async function ssrTransformScript(
 
     // export * from './foo'
     if (node.type === 'ExportAllDeclaration') {
+      s.remove(node.start, node.end)
+      const importId = defineImport(node, node.source.value as string)
       if (node.exported) {
-        const importId = defineImport(node, node.source.value as string)
-        s.remove(node.start, node.end)
         defineExport(node.end, node.exported.name, `${importId}`)
       } else {
-        const importId = defineImport(node, node.source.value as string)
-        s.remove(node.start, node.end)
         s.appendLeft(node.end, `${ssrExportAllKey}(${importId});`)
       }
     }