+++ playground/html/vite.config.js
@@ -1,4 +1,4 @@
-const { resolve } = require('path')
+const { resolve } = require('node:path')
 
 /**
  * @type {import('vite').UserConfig}