+++ playground/ssr-deps/require-absolute/index.js
@@ -1,3 +1,3 @@
-const path = require('path')
+const path = require('node:path')
 
 module.exports.hello = () => require(path.resolve(__dirname, './foo.js')).hello